:art: finder 페이지 ui 폴더 구조 변경 #69
