+++ src/pages/play/finder/ui/SearchLobby/SearchModalContent.tsx
@@ -1,40 +1,40 @@
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 
-import { postEnterRoom } from '../api';
-import * as styles from './index.css';
-import MessageModalContent from './MessageModalContent';
-
-import type { RequestEnterRoom } from '@/types/play';
+import { postEnterRoom } from '../../api';
+import * as styles from '../index.css';
+import MessageModalContent from '../MessageModalContent';
 
 import ColoredButton from '@/components/button/ColoredButton';
 import Input from '@/components/input';
-
 
 import { ROUTE } from '@/constants/routes';
 
 interface Props {
   closeModal: () => void;
-  roomCode: string;
 }
 
-const PasswordModal = ({ roomCode, closeModal }: Props) => {
+const SearchModalContent = ({ closeModal }: Props) => {
   const navigate = useNavigate();
   const [isValid, setIsValid] = useState(true);
   const [roomInfo, setRoomInfo] = useState<RequestEnterRoom>({
-    roomCode,
+    roomCode: '',
     roomPassword: '',
   });
 
   const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
     const { name, value } = event.target;
+    if (name === 'roomCode' && value.length > 4) return;
     if (name === 'roomPassword' && value.length > 4) return;
-    setRoomInfo((info) => ({ ...info, [name]: value }));
+    setRoomInfo((info) => ({
+      ...info,
+      [name]: value,
+    }));
   };
 
   const handleClickButton = async () => {
     const roomId = await postEnterRoom(roomInfo);
-    if (!roomId || roomInfo.roomPassword.length !== 4) {
+    if (!roomId) {
       setIsValid(false);
     } else {
       setIsValid(true);
@@ -45,7 +45,7 @@ const PasswordModal = ({ roomCode, closeModal }: Props) => {
   if (!isValid) {
     return (
       <>
-        <MessageModalContent failed='PASSWORD' />
+        <MessageModalContent failed='SEARCH' />
         <div className={styles.modalTwoButtonWrapper}>
           <ColoredButton
             size='small'
@@ -59,16 +59,24 @@ const PasswordModal = ({ roomCode, closeModal }: Props) => {
       </>
     );
   }
+
   return (
     <div className={styles.contentBox}>
-      <span className={styles.text}>이 방의 Colored들은 </span>
-      <span className={styles.text}>아무하고나 쉽게 싸우지 않습니다 </span>
+      <span>방장한테 허락은 맡으셨죠?</span>
+      <Input
+        name='roomCode'
+        placeholder='대기실 코드 4자리를 입력하세요'
+        size='small'
+        type='text'
+        value={roomInfo.roomCode}
+        onChange={handleChange}
+      />
       <Input
         name='roomPassword'
         placeholder='비밀번호 4자리를 입력하세요'
         size='small'
         type='text'
-        value={roomInfo?.roomPassword}
+        value={roomInfo.roomPassword}
         onChange={handleChange}
       />
       <div className={styles.modalTwoButtonWrapper}>
@@ -88,4 +96,5 @@ const PasswordModal = ({ roomCode, closeModal }: Props) => {
     </div>
   );
 };
-export default PasswordModal;
+
+export default SearchModalContent;
